style(cli): move repl to a sub directory
