+++ crates/jstz_cli/src/main.rs
@@ -4,9 +4,7 @@ use clap::Parser;
 mod account;
 mod bridge;
 mod config;
-mod debug_api;
 mod deploy;
-mod js_logger;
 mod jstz;
 mod kv;
 mod logs;
